Copied Over Code from User Component and Service
Copied over the code from the user-list component and the user-service to be used in the todo component and service. Made all the necessary changes to create a smooth transition between the code for the users and the code for the todos. All coding of this commit was done in the typescript.